Move asset status partition cache methods into a test class
Summary:
This allows us to swap out the instance in different contexts while still preserving the core test logic.

Test Plan: BK
